Ports basic_test.go to Windows
This commit ports the existing TestVolumeLifecycle test in basic_test.go to
Windows.

* Enables spawning of a microsoft/nanoserver Container on a Windows host.
* Normalizes volume name in GetUniqueVolumeName based on the test platform.
