+++ tests/constants/dockercli/cmddefault.go
@@ -1,0 +1,30 @@
+// Copyright 2017 VMware, Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Default constants related to docker cli.
+
+// +build !winutil
+
+package dockercli
+
+const (
+	// ContainerImage busybox container image
+	ContainerImage = " busybox "
+
+	// TestContainer test busybox container that keeps running
+	TestContainer = ContainerImage + " tail -f /dev/null "
+
+	// ContainerMountPoint mount point where a volume is mounted inside a container
+	ContainerMountPoint = "/vol"
+)
